Moved to the core module
